requirements.txt
Fastapi
bcrypt==4.0.1
passlib[bcrypt]
python-jose